services/llm-extraction/requirements.txt
fastapi>=0.128.6
httpx>=0.27
langchain==1.2.10
langchain-openai==1.1.9
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv==1.2.1
uvicorn>=0.40.0
PyYAML
pandas==3.0.1
